No warnings

+++ pg_sparse/src/api/mod.rs
@@ -9,10 +9,10 @@ pub fn sparse_cosine_distance(left: Sparse, right: Sparse) -> f32 {
     let mut right_map = HashMap::new();
 
     for entry in &left.entries {
-        left_map.insert(entry.0, entry.1 as f32);
+        left_map.insert(entry.0, entry.1);
     }
     for entry in &right.entries {
-        right_map.insert(entry.0, entry.1 as f32);
+        right_map.insert(entry.0, entry.1);
     }
 
     let max_length = left.n.max(right.n);
@@ -22,8 +22,8 @@ pub fn sparse_cosine_distance(left: Sparse, right: Sparse) -> f32 {
     let mut right_norm: f32 = 0.0;
 
     for position in 0..max_length {
-        let left_value = *left_map.get(&(position + 1)).unwrap_or(&0.0) as f32;
-        let right_value = *right_map.get(&(position + 1)).unwrap_or(&0.0) as f32;
+        let left_value = *left_map.get(&(position + 1)).unwrap_or(&0.0);
+        let right_value = *right_map.get(&(position + 1)).unwrap_or(&0.0);
 
         dot_product += left_value * right_value;
         left_norm += left_value.powi(2);
@@ -34,7 +34,7 @@ pub fn sparse_cosine_distance(left: Sparse, right: Sparse) -> f32 {
         return -1.0;
     }
 
-    (dot_product / (left_norm.sqrt() * right_norm.sqrt())) as f32
+    dot_product / (left_norm.sqrt() * right_norm.sqrt())
 }
 
 extension_sql!(
